+++ src/Smartstore.Modules/Smartstore.PayPal/Components/PayPalSepaViewComponent.cs
@@ -1,52 +1,25 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Smartstore.Web.Components;
 
 namespace Smartstore.PayPal.Components
 {
-    public class PayPalSepaViewComponent : SmartViewComponent
+    /// <summary>
+    /// Renders PayPal button widget (funding source: sepa).
+    /// </summary>
+    public class PayPalSepaViewComponent : PayPalViewComponentBase
     {
-        private readonly PayPalSettings _settings;
-
-        public PayPalSepaViewComponent(PayPalSettings settings)
+        protected override IViewComponentResult InvokeCore()
         {
-            _settings = settings;
-        }
-
-        /// <summary>
-        /// Renders PayPal buttons widget.
-        /// </summary>
-        /// <param name="isPaymentInfoInvoker">Defines whether the widget is invoked from payment method's GetPaymentInfoWidget.</param>
-        /// <param name="isSelected">Defines whether the payment method is selected on page load.</param>
-        public IViewComponentResult Invoke(string funding, bool isPaymentInfoInvoker, bool isSelected)
-        {
-            // If client id or secret haven't been configured yet, don't render buttons.
-            if (!_settings.ClientId.HasValue() || !_settings.Secret.HasValue())
-            {
-                return Empty();
-            }
-
-            var routeIdent = Request.RouteValues.GenerateRouteIdentifier();
-            var isPaymentSelectionPage = routeIdent == "Checkout.PaymentMethod";
-
-            if (isPaymentSelectionPage && isPaymentInfoInvoker)
-            {
-                return Empty();
-            }
-
             // Get displayable options from settings depending on location (OffCanvasCart or Cart).
-            var isCartPage = routeIdent == "ShoppingCart.Cart";
-            if (isCartPage && !_settings.ShowButtonOnCartPage)
+            var isCartPage = RouteIdent == "ShoppingCart.Cart";
+            if (isCartPage && !Settings.FundingsCart.Contains(FundingOptions.sepa.ToString()))
             {
                 return Empty();
             }
 
             var model = new PublicPaymentMethodModel
             {
-                IsPaymentSelection = isPaymentSelectionPage,
-                ButtonColor = _settings.ButtonColor,
-                ButtonShape = _settings.ButtonShape,
-                IsSelectedMethod = isSelected,
-                Funding = funding
+                ButtonColor = Settings.ButtonColor,
+                ButtonShape = Settings.ButtonShape
             };
 
             return View(model);