+++ src/Smartstore.Modules/Smartstore.PayPal/Filters/OffCanvasShoppingCartFilter.cs
@@ -24,29 +24,35 @@ namespace Smartstore.PayPal.Filters
 
         public async Task OnResultExecutionAsync(ResultExecutingContext filterContext, ResultExecutionDelegate next)
         {
-            if (!await _payPalHelper.IsPayPalStandardActiveAsync())
-            {
-                await next();
-                return;
-            }
-            
             // If client id or secret haven't been configured yet, don't show button.
             if (!_settings.ClientId.HasValue() || !_settings.Secret.HasValue())
             {
                 await next();
                 return;
             }
-
-            if (!_settings.ShowButtonInMiniShoppingCart)
-            {
-                await next();
-                return;
-            }
-
+            
             // Should only run on a full view rendering result or HTML ContentResult.
             if (filterContext.Result is StatusCodeResult || filterContext.Result.IsHtmlViewResult())
             {
-                _widgetProvider.RegisterViewComponent<PayPalViewComponent>("offcanvas_cart_summary");
+                var fundings = _settings.FundingsOffCanvasCart;
+
+                // PayPalStandard
+                if (fundings.Contains(FundingOptions.paypal.ToString()) && await _payPalHelper.IsPayPalStandardActiveAsync())
+                {
+                    _widgetProvider.RegisterViewComponent<PayPalViewComponent>("offcanvas_cart_summary");
+                }
+
+                // SEPA
+                if (fundings.Contains(FundingOptions.sepa.ToString()) && await _payPalHelper.IsSepaActiveAsync())
+                {
+                    _widgetProvider.RegisterViewComponent<PayPalSepaViewComponent>("offcanvas_cart_summary");
+                }
+
+                // PayLater
+                if (fundings.Contains(FundingOptions.paylater.ToString()) && await _payPalHelper.IsPayLaterActiveAsync())
+                {
+                    _widgetProvider.RegisterViewComponent<PayPalPayLaterViewComponent>("offcanvas_cart_summary");
+                }
             }
 
             await next();